requirements.txt
mcp[cli]>=1.27.0
numpy>=1.26.0
pandas>=2.2.0
pydantic>=2.7.0
pytest>=8.0.0
python-dotenv>=1.0.1
requests>=2.32.0
starlette>=0.37.2
uvicorn>=0.30.0

